Create Enforce DevFund Commitments with a Legal Charter.rst
This was written up by @mistfpga, but this is 100% lex-node's proposal.

whist there is/was some controversy around 2-2, the trademark and signoff - this proposal should not be stopped by those, @lex-node has expressed strong desire to continue with this, he just did not have the time to write it up. It will be adapted to fit whatever format happens to be going forward. I have taken some liberty when addressing comments, lex-node seemed happy with my answers.