ZIP 1008: Fund ECC for Two More Years
This was proposed by user kek on the Zcash forums. They are unable to continue
with this proposal due to real life stuff. They have asked me to act as advocate.
I am pushing this ZIP forward until kek gets time and would like to come back and
take over.

Signed-off-by: Daira Hopwood <daira@electriccoin.co>