Create delayed_choice.rst
This was proposed by user KEK on the zcash forums. They are unable to continue with this proposal due to real life stuff. They have asked me to act as advocate I am pushing this zip forward until kek gets time and would like to come back and take over