Create Changing Zcash supply curve.rst
This was proposed by KEK on the zcash forums. They are unable to continue with this proposal due to real life stuff. They have asked me to act as advocate I am pushing this zip forward until kek gets time and would like to come back and take over.  This proposal is very sparse. However kek is a very involved member of the community and I expect him to be back online and able to flesh this out more.  Until such a time, it will have to stay in this state.  Written up by mistfpga.