Remove Howard Loo from credits section
I request that my name please be removed from this ZIP. I do not support it and as a member of the Community Panel I voted against it. I merely provided @tromer with some brief feedback one evening about the copyright portions of the ZIP. I did give permission for my name to be mentioned publicly as having provided feedback, but I did not realize that my name would be part of the actual ZIP text. Thank you for considering my request.