Remove link to the "GNU Kind Communication Guidelines" given Richard Stallman's involvement (see https://rms-open-letter.github.io/appendix), and that document's stance on pronouns.
In particular, although the document has changed since the open letter's reference to it, its
current version says that gender-neutral pronouns "don't conflict with any possible gender
identity". That is incorrect, and not compatible with our CoC. Always use a person's stated
pronouns if they are known; use gender-neutral pronouns only when the correct ones are unknown
or when not referring to a specific person.

Besides the pronoun issue and the association with RMS, a universal exhortation to "assume good
faith" is verging on tone-policing when applied to marginalized communities. In any case,
ZIP 0 is not about that topic and the link is out-of-place.

Signed-off-by: Daira Hopwood <daira@jacaranda.org>