ZIP 0: Remove reference to the "GNU Kind Communication Guidelines" given Richard Stallman's involvement (see https://rms-open-letter.github.io/appendix).
I thought I had already done this in db3b2f72d4fe70365ca4f9cd0e618a8635c835cf.

A universal exhortation to "assume good faith" is verging on tone-policing
when applied to marginalized communities. In any case, ZIP 0 is not about
that topic and the link is out-of-place.

Signed-off-by: Daira-Emma Hopwood <daira@jacaranda.org>