Changes to the issuance authorization signature and issuance validating key (#107)
This makes changes to the issuance authorization signature and issuance
validating key fields of the V6 transaction format in order to make them
length-prefixed encodings.

Further, it also adds a version byte to the start of both the above
fields (as part of the definition of the scheme, not part of the
transaction format).