Ensure that the protocol spec is rendered when publishing.
Publishing of the rendered protocol spec was broken by the last commit of
#1120, which changed the behaviour of the `render` action to render only
ZIPs. It wasn't noticed because files that are no longer present in the
rendering won't be automatically deleted from the `publish` branch (as
currently intended, but see #1140), and that was the case for the protocol
spec PDFs.

Signed-off-by: Daira-Emma Hopwood <daira@jacaranda.org>