CI: Ensure that `origin/main` is available alongside `origin/publish`
The only configuration option that ensures branch refs other than the
explicitly-specified one are present, is `fetch-depth: 0`. Unfortunately
a side-effect of this is that we check out the entire history of the
repository including all past binary versions of the protocol spec PDF,
which will slow down the render workflow, but until we come up with a
different approach for performing the merge from `origin/main` to
`origin/publish`, we need to bear the performance hit.